dynamically import ParallaxDaoStats to fix runtime error
