+++ src/components/Governance/ParallaxDaoStats/ParallaxDaoStats.tsx
@@ -1,0 +1,12 @@
+import ParallaxDaoStatsElement from '@/components/Governance/ParallaxDaoStats/ParallaxDaoStatsElement'
+import ParallaxText, { type ParallaxTextProps } from '@/components/common/ParallaxText'
+
+const ParallaxDaoStats = (props: ParallaxTextProps) => {
+  return (
+    <ParallaxText {...props}>
+      <ParallaxDaoStatsElement items={props.items} />
+    </ParallaxText>
+  )
+}
+
+export default ParallaxDaoStats
